Refactor: PostCard, TagCard 리팩토링
